Added Business Exception Added Fluent Validation Added Sql server and install the packages

+++ HRLeaveManagementApplication/Features/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using HRLeaveManagement.Domain;
 using HRLeaveManagementApplication.Contracts.DataAccess;
+using HRLeaveManagementApplication.Exceptions;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -13,14 +14,19 @@ namespace HRLeaveManagementApplication.Features.Commands.DeleteLeaveType
     {
         private readonly ILeaveTypeRepository _leaveTypeRepository;
 
-        public DeleteLeaveTypeCommandHandler(ILeaveTypeRepository leaveTypeRepository)=>  _leaveTypeRepository = leaveTypeRepository;
-        
+        public DeleteLeaveTypeCommandHandler(ILeaveTypeRepository leaveTypeRepository) => _leaveTypeRepository = leaveTypeRepository;
+
         public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
         {
-            
+
             //retrieve to domain entity object
             var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
+
             //verify that record exists.
+            if (leaveTypeToDelete == null)
+            {
+                throw new NotFoundException(nameof(LeaveType), request.Id);
+            }
 
             //remove from the database
             await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
